Move the content process handling to a separate file in TermControl project
