+++ src/cascadia/TerminalControl/TerminalControlLib.vcxproj
@@ -101,6 +101,9 @@
     <ClCompile Include="TermControl.cpp">
       <DependentUpon>TermControl.xaml</DependentUpon>
     </ClCompile>
+    <ClCompile Include="TermControlContentManagement.cpp">
+      <DependentUpon>TermControl.xaml</DependentUpon>
+    </ClCompile>
     <ClCompile Include="TSFInputControl.cpp">
       <DependentUpon>TSFInputControl.xaml</DependentUpon>
     </ClCompile>
